+++ changelog.md
@@ -7,6 +7,7 @@
 ## Fixes:
 * **[Mission Generator]** Fix mission generation error when there are too many radio frequency to setup for the Mig-21
 * **[Mission Generator]** Fix ground units not moving forward
+* **[Flight Planner]** Fix creation of custom waypoints.
 
 # 2.3.1